melhorando a api

+++ src/main/java/com/portfolio/api/Controllers/ProjetoController.java
@@ -2,6 +2,7 @@ package com.portfolio.api.Controllers;
 
 import com.portfolio.api.Entities.Projeto;
 import com.portfolio.api.Repositories.ProjetoRepository;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -16,22 +17,38 @@ public class ProjetoController {
     private ProjetoRepository repository;
 
     @GetMapping("/api/projetos")
-    public List<Projeto> getAll() {
-        return repository.findAll();
+    public ResponseEntity<List<Projeto>> getAll() {
+        var Projetos = repository.findAll();
+        return ResponseEntity.ok(Projetos);
     }
 
     @PostMapping("/api/projetos")
+    @Transactional
     public Projeto create(@RequestBody Projeto projeto) {
         return repository.save(projeto);
     }
 
     @GetMapping("/api/projetos/{id}")
-    public Projeto show(@PathVariable Long id) {
-        return repository.findById(id).orElse(null);
+    public ResponseEntity<Projeto> show(@PathVariable Long id) {
+        var projeto = repository.findById(id).orElse(null);
+
+        if(projeto == null) {
+            return ResponseEntity.notFound().build();
+        }
+
+        return ResponseEntity.ok(projeto);
     }
 
     @DeleteMapping("/api/projetos/{id}")
-    public void delete(@PathVariable Long id) {
-        repository.deleteById(id);
+    @Transactional
+    public ResponseEntity delete(@PathVariable Long id) {
+        var projeto = repository.findById(id);
+
+        if(projeto.isPresent()) {
+            repository.deleteById(id);
+            return ResponseEntity.noContent().build();
+        }
+
+        return ResponseEntity.notFound().build();
     }
 }
